refactor: [logging] pull out logging setup from the main function
Simplify the cli main function by pulling out the logging setup. The cli
function is more difficualt to test, the more we can full out the easier
it is to test.
