+++ src/dvsim/utils/__init__.py
@@ -48,6 +48,3 @@ __all__ = (
     "run_cmd_with_timeout",
     "subst_wildcards",
 )
-
-# For verbose logging
-VERBOSE = 15
